+++ packages/python-client/lmk/jupyter/constants.py
@@ -1,5 +1,3 @@
-from lmk.constants import VERSION
-
 MODULE_NAME = "@lmkapp/jupyter-widget"
 
-MODULE_VERSION = VERSION
+MODULE_VERSION = "1.0.3-dev0"